Fixed observatory tabs, tranform many pages in one with components and shared index tabs, closed #134
