+++ components/Common/state.js
@@ -1,0 +1,3 @@
+import { ref } from 'vue';
+
+export const activeTabIndex = ref(0); // Shared state for the active tab index
