+++ components/Header/HeaderMenu/subMenuEntriesObservatory.js
@@ -1,36 +1,12 @@
 // Displayed in Header and Footer
 
-const subMenuEntriesObservatory = [
-  {
-    title: "Trends",
-    title_hide: "Trends",
-    to: "/observatory/Trends",
-    icon: "mdi-chart-line",
-  },
-  {
-    title: "FAIRness Scoreboard",
-    title_hide: "FAIRness Scoreboard",
-    to: "/observatory/FAIRness",
-    icon: "mdi-bullseye",
-  },
-  {
-    title: "FAIR Evaluator",
-    title_hide: "FAIR Evaluator",
-    to: "/observatory/Evaluation",
-    icon: "mdi-check-circle-outline",
-  },
-  {
-    title: "Data",
-    title_hide: "Observatory Data",
-    to: "/observatory/Data",
-    icon: "mdi-database",
-  },
-  {
-    title: "About",
-    title_hide: "Observatory About",
-    to: "/observatory/About",
-    icon: "mdi-information",
-  },
+const tabsItems = [
+  { label: "Trends", slot: "trends", icon: "i-charm-chart-line", path: "observatory/", index: 0},
+  { label: "FAIRness Scoreboard", slot: "fairness", icon: "i-charm-disc", path: "observatory/", index: 1 },
+  { label: "FAIRsoft Evaluator", slot: "fairsoft", icon: "i-cil-check-circle", path: "observatory/", index: 2 },
+  { label: "Data", slot: "data", icon: "i-dashicons-database", path: "observatory/", index: 3 },
+  { label: "About", slot: "about", icon: "i-zondicons-information-solid", path: "observatory/", index: 4 },
 ];
 
-export default subMenuEntriesObservatory;
+export default tabsItems;
+
