--- components/Observatory/TabsPlotInfo.vue
@@ -1,6 +1,0 @@
-<template>
-
-</template>
-<script setup lang="ts">
-
-</script>
