+++ layouts/default.vue
@@ -17,7 +17,6 @@
 
 <script setup lang="ts">
 import HeaderMenu from "~/components/Header/HeaderMenu.vue";
-import BreadcrumbsBar from "~/components/Common/BreadcrumbsBar.vue";
 import ScrollToTop from '~/components/Common/ArrowUp.vue';
 import TheFooter from "~/components/TheFooter.vue";
 // useHead({
