--- pages/observatory/PostGitHubInstall.vue
@@ -1,6 +1,0 @@
-<template>
-    <h1>PostGitHubInstall</h1>
-</template>
-
-<script setup lang="ts">
-</script>
